#KJ-54 verify the correctness of findQualifierForName(). Verified and it is correct. From method documentation:
Computes into an expression of the form elabRes(cast(QualClass, 'QThis(QualClass))),
where QualClass is searched in the enclosing context of the first argument,
being the innermost class that have a method with name = the second argument.
Or noValue if no suitable result is found.

This algorithm is correct even when both inner and outer class have methods with the same name.
Java don't support method overloading in the context of nesting. In other words, inner class methods
don't overload but rather hide methods with the same name from outer classes. See program
inner_in_81_inner_overl_outer as example.